move HN structs in `client` module to a new module `model`
